+++ shape/osci_CubicBezierCurve.h
@@ -6,16 +6,16 @@
 namespace osci {
 class CubicBezierCurve : public Shape {
 public:
-	CubicBezierCurve(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4);
+	CubicBezierCurve(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4);
 
-	Point nextVector(double drawingProgress) override;
-	void scale(double x, double y, double z) override;
-	void translate(double x, double y, double z) override;
-	double length() override;
+	Point nextVector(float drawingProgress) override;
+	void scale(float x, float y, float z) override;
+	void translate(float x, float y, float z) override;
+	float length() override;
 	std::unique_ptr<Shape> clone() override;
 	std::string type() override;
 
-	double x1, y1, x2, y2, x3, y3, x4, y4;
+	float x1, y1, x2, y2, x3, y3, x4, y4;
 	
 };
 } // namespace osci
