+++ shape/osci_Line.h
@@ -6,20 +6,20 @@
 namespace osci {
 class Line : public Shape {
 public:
-	Line(double x1, double y1, double x2, double y2);
-	Line(double x1, double y1, double z1, double x2, double y2, double z2);
+	Line(float x1, float y1, float x2, float y2);
+	Line(float x1, float y1, float z1, float x2, float y2, float z2);
 	Line(Point p1, Point p2);
 
-	Point nextVector(double drawingProgress) override;
-	void scale(double x, double y, double z) override;
-	void translate(double x, double y, double z) override;
-	static double length(double x1, double y1, double z1, double x2, double y2, double z2);
-	double length() override;
+	Point nextVector(float drawingProgress) override;
+	void scale(float x, float y, float z) override;
+	void translate(float x, float y, float z) override;
+	static float length(float x1, float y1, float z1, float x2, float y2, float z2);
+	float length() override;
 	std::unique_ptr<Shape> clone() override;
 	std::string type() override;
 	Line& operator=(const Line& other);
 
-	double x1, y1, z1, x2, y2, z2;
+	float x1, y1, z1, x2, y2, z2;
 	
 };
 } // namespace osci
